Preparando estructura de archivos frontend
