+++ client/src/App.jsx
@@ -1,0 +1,13 @@
+import {BrowserRouter, Route, Routes} from 'react-router-dom'
+
+function App() {
+  return (
+    <BrowserRouter>
+      <Routes>
+        
+      </Routes>
+    </BrowserRouter>
+  );
+}
+
+export default App;
